+++ Sources/ChessKitEngineCore/Engines/Extensions/lc0+engine.cpp
@@ -5,29 +5,13 @@
 
 #include "lc0+engine.h"
 
-#include "../lc0/src/chess/board.h"
-#include "../lc0/src/engine.h"
-
-using namespace lczero;
-
-EngineLoop loop;
+#include "../lc0/src/_main.h"
 
 void Lc0Engine::initialize() {
-    InitializeMagicBitboards();
-    loop.Initialize();
-    loop.RunLoop();
+    const char* argv[] = { "uci" };
+    _main(sizeof(argv) / sizeof(argv[0]), argv);
 }
 
 void Lc0Engine::deinitialize() {
-    
-}
 
-void Lc0Engine::send_command(const std::string &cmd) {
-    auto command = loop.ParseCommand(cmd);
-    
-    try {
-        loop.DispatchCommand(command.first, command.second);
-    } catch(std::exception& e) {
-        // ignore unsupported commands
-    }
 }